scrapling/requirements.txt
flask==3.0.0
requests==2.32.5
beautifulsoup4==4.14.3
openpyxl==3.1.5
lxml==5.3.0
scrapling==0.3.3
# playwright==1.49.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyee==12.0.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
google-generativeai==0.8.4
python-dotenv==1.0.1
